requirements.txt
-f git+https://github.com/facebookresearch/detectron2.git
fvcore==0.1.5.post20221221
gradio==3.38.0
iopath==0.1.9
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.8.0.74
Pillow
scikit_learn
tabulate==0.9.0
torch==1.13.0
torchvision==0.14.0
